Add TLS as setting instead of depending on port 443
