+++ dynamodb/src/main/scala/akka/stream/alpakka/dynamodb/DynamoSettings.scala
@@ -12,6 +12,7 @@ final class DynamoSettings private (
     val region: String,
     val host: String,
     val port: Int,
+    val tls: Boolean,
     val parallelism: Int,
     val credentialsProvider: com.amazonaws.auth.AWSCredentialsProvider
 ) extends AwsClientSettings {
@@ -21,7 +22,12 @@ final class DynamoSettings private (
 
   def withRegion(value: String): DynamoSettings = copy(region = value)
   def withHost(value: String): DynamoSettings = copy(host = value)
-  def withPort(value: Int): DynamoSettings = copy(port = value)
+  def withPort(value: Int): DynamoSettings =
+    if (value == 443) {
+      copy(port = value, tls = true)
+    } else copy(port = value)
+  def withTls(value: Boolean): DynamoSettings =
+    if (value == tls) this else copy(tls = value)
   def withParallelism(value: Int): DynamoSettings = copy(parallelism = value)
   def withCredentialsProvider(value: com.amazonaws.auth.AWSCredentialsProvider): DynamoSettings =
     copy(credentialsProvider = value)
@@ -30,12 +36,14 @@ final class DynamoSettings private (
       region: String = region,
       host: String = host,
       port: Int = port,
+      tls: Boolean = tls,
       parallelism: Int = parallelism,
       credentialsProvider: com.amazonaws.auth.AWSCredentialsProvider = credentialsProvider
   ): DynamoSettings = new DynamoSettings(
     region = region,
     host = host,
     port = port,
+    tls = tls,
     parallelism = parallelism,
     credentialsProvider = credentialsProvider
   )
@@ -60,6 +68,9 @@ object DynamoSettings {
     val region = c.getString("region")
     val host = c.getString("host")
     val port = c.getInt("port")
+    val tls =
+      if (c.hasPath("tls")) c.getBoolean("tls")
+      else port == 443
     val parallelism = c.getInt("parallelism")
     val awsCredentialsProvider = {
       if (c.hasPath("credentials.access-key-id") &&
@@ -73,6 +84,7 @@ object DynamoSettings {
       region,
       host,
       port,
+      tls,
       parallelism,
       awsCredentialsProvider
     )
@@ -96,8 +108,9 @@ object DynamoSettings {
   ): DynamoSettings = new DynamoSettings(
     region,
     host,
-    443,
-    4,
+    port = 443,
+    tls = true,
+    parallelism = 4,
     new DefaultAWSCredentialsProviderChain()
   )
